added module level documentation to 'classes'
Typically, there are two ways to add documentation to a module.

The first method (which is more idiomatic) is two write a top-level
documentation block at the top of the module source file as such:

a/b/c.rs:
    //! doc for module c

The other method is to adorn a `mod` item with a documentation comment,
as such:

a/b.rs
    /// doc for module c
    mod c;

The reason I preffered to go for the second option is that the first
option is impossible due to the classes.rs file being overwritten. I
could have hardcoded the documentation into the
`lib/RustClientGenerator` script but that would have been a considerable
worse solution.

Signed-off-by: Ben Raz <ben.raz2008@gmail.com>