removed _trait_buff
The original idea was to have all traits delegated to _trait_buff and be
saved to a seperate module from the classes. Ideally, a base class A
would be defined in the following way:

classes.rs:
    pub struct A {}

traits.rs:
    pub trait A {}

Unfortunately, this would have resulted in an ugly situation when it
comes to namespace pollution. If one wanted to use classes::*, he would
now be forbidden from using traits::*, and vice versa. It is true that
glob imports are a bad practice anyway but that does not mean that they
are not common and it makes no sense to build the api in a way that
needlessly makes impossible that use case, when it easier to build the
API in a different way. Namely, for a given class A, classes.rs will
contain the following:

    pub struct A {}

    pub trait IA {}

Signed-off-by: Ben Raz <ben.raz2008@gmail.com>