introduced prefixWithTab function
prefixing with tabs was previously handled by the functions that format
rust items not the ones that call them and use the string
representation. This has been changed. Now the functions getting called
give unindented output and it is up to the function calling them to
prefix them with tabs.

Signed-off-by: Ben Raz <ben.raz2008@gmail.com>