converted all implicit blocks to explicit blocks
PHP allows a shorthand 'implicit block' syntax where the following
statement:

    if (a) {
	do_b();
    }

can be written in shorthanded form:

    if (a) do_b();

or:

    if (a)
	do_b();

This rule only applies if there is one statement in the block. If one
added another statement, we would have to add back the braces. This
unnecessarily complicates the diff, so I made the decision to convert
all existing usages of this syntax to the explicit form, even though the
implicit form looks nicer.

Signed-off-by: Ben Raz <ben.raz2008@gmail.com>